move pkg/controller/cainjector to cmd/cainjector/controller
